+++ src/main/java/com/management/service/AdminService.java
@@ -97,4 +97,13 @@ public interface AdminService {
      * @Date: 2018/12/27
      */
     Result findExpertList();
+
+    /**
+     * @Description: 通过项目id找到这个项目所有的审核专家对项目的审核状况
+     * @Param: [projectId]
+     * @Return: com.management.model.ov.Result
+     * @Author: ggmr
+     * @Date: 18-8-1
+     */
+    Result expertOpinionList(int projectId);
 }
